Added placeSelected signal
The placeSelected signal emits the model item that was selected,
providing access to the item for all of its listeners.
This allows us to share the item to components that need to display
the item's information when its selected.

Renamed select and deselect functions to selectMarker and deselectMarker
to be more explicit